requirements.txt
streamlit
tensorflow
opencv-python-headless
numpy
scikit-learn
matplotlib
Pillow
